requirements.txt
   streamlit
   pillow
   numpy
   opencv-python
   scikit-learn
   joblib